+++ CHANGELOG.md
@@ -16,11 +16,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - Share server host via cookies
 - Generate a patient code
+- Display of Record number according to preferences
 - List of result (no savings result)
 - List records : action to enter result
 - library python-barcode
 - library pdfkit
 - footer page php and python same number
+- contributors : url site labbook changed
 
 ## [2.9.0-beta.1] - 2020-03-20
 ### Added